Q4_K_F: 2nd shot at Metal
Now beating Q4_0 by a small margin: 21.1 vs 21.5 ms/token
at 7B and 37.48 vs 38.13 at 13B.

If I put return as the 1st statement in the dot kernel, I
measure 8.5 ms/token run time. So, kernel run time is
12.6 ms, so we get ~303.6 GB/s memory throughput.

Is this as good as it gets on M2 Max?
I would be curious to have somebody run it on the
38-core GPU version.